feat: Use esp_new_jpeg intead of esp_jpeg. Refactor jpeg decode/encode functions.

+++ esp-dl/vision/image/dl_image_preprocessor.cpp
@@ -14,8 +14,8 @@ ImagePreprocessor::ImagePreprocessor(Model *model,
     assert(m_model_input->dtype == DATA_TYPE_INT8 || m_model_input->dtype == DATA_TYPE_INT16);
     assert(m_model_input->shape[3] == m_mean.size() && m_mean.size() == m_std.size());
     m_output = {.data = m_model_input->data,
-                .width = m_model_input->shape[2],
-                .height = m_model_input->shape[1],
+                .width = (uint16_t)m_model_input->shape[2],
+                .height = (uint16_t)m_model_input->shape[1],
                 .pix_type = (m_model_input->dtype == DATA_TYPE_INT8) ? DL_IMAGE_PIX_TYPE_RGB888_QINT8
                                                                      : DL_IMAGE_PIX_TYPE_RGB888_QINT16};
     if (m_model_input->dtype == DATA_TYPE_INT8) {
